requirements.txt
streamlit
pandas
matplotlib
textblob
openai
python-dotenv
